+++ src/item_location.cpp
@@ -668,7 +668,7 @@ void item_location::deserialize( JsonIn &js )
     } else if( type == "vehicle" ) {
         vehicle *const veh = veh_pointer_or_null( g->m.veh_at( pos ) );
         int part = obj.get_int( "part" );
-        if( veh && part >= 0 && part < static_cast<int>( veh->part_count() ) ) {
+        if( veh && part >= 0 && part < veh->part_count() ) {
             ptr.reset( new impl::item_on_vehicle( vehicle_cursor( *veh, part ), idx ) );
         }
     } else if( type == "in_container" ) {
